Add linux metrics

+++ Mnemox.Machine.Metrics/ICpuMetrics.cs
@@ -2,7 +2,7 @@
 {
     public interface ICpuMetrics
     {
-        float GetCpuUsagePercentage();
-        float GetCurrentProcessCpuUsagePercentage();
+        double GetCpuUsagePercentage();
+        double GetCurrentProcessCpuUsagePercentage();
     }
 }
