+++ Mnemox.Machine.Metrics/Structures/OsMetrics.cs
@@ -2,7 +2,7 @@
 {
     public class OsMetrics
     {
-        public float TotalCpuUsagePercentage { get; set; }
+        public double TotalCpuUsagePercentage { get; set; }
         public ulong CapacityBytes { get; set; }
     }
 }
